% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillobenchmark-package.R
\docType{package}
\name{armadillobenchmark-package}
\alias{armadillobenchmark}
\alias{armadillobenchmark-package}
\title{armadillobenchmark: Adapted ATT benchmark for Armadillo}
\description{
Adaptation of the ATT benchmark for variable matrix or vector size in Armadillo (C++).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/cpp11eigen}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11eigen/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
